+++ tests/integration/ConfigKeys.php
@@ -6,6 +6,6 @@ class ConfigKeys
 {
     public function loadConfigs()
     {
-        return parse_ini_file(__DIR__ . '..\..\..\.env');
+        return parse_ini_file(__DIR__ . '../../../.env');
     }
 }