+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## V1.5.5
+### Added
+- Integration tests for multipayment scenario
+
+### Changed
+- Modify composer.json in autoload-dev to load all class in /tests
+
 ## V1.5.4
 ### Added
 - Adds a step to the script that runs via GitHub Actions to create a version tag for the internal repository